+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `/collections`, `/divisions`, `/divisions/[slug]`, and `/collections/lane/[slug]` pages (DR-3021)
 - Added middleware to redirect unpublished pages (DR-3021)
 - Added `PageLayout` component (DR-3022)
-
+- Added `/items` pages (DR-3029)
 
 ## [0.1.8] 2024-06-06